CS/import improvements
