+++ mpf/file_interfaces/__init__.py
@@ -1,5 +1,1 @@
 """Contains config file interfaces."""
-__all__ = ('yaml_interface',
-           # 'json_interface',  will uncomment when ready
-           # 'xml_interface',
-           )
